reorganize and clean
